pyproject.toml
[tool.poetry]
name = "fastapi-gradio-starter"
version = "0.1.0"
description = ""
authors = ["ren-surf <ren.nakamura.engineer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
gradio = "^5.9.0"
python-dotenv = "^1.0.1"
ollama = "^0.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
  "C9",
  "E",
  "F",
  "W",
  "I",
]

[tool.mypy]
python_version = "3.11"
show_error_context = true
show_column_numbers = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true